feat: apply elegant dark theme and update admin membership flow

+++ app/controllers/api/v1/admin/categories_controller.rb
@@ -1,10 +1,12 @@
 class Api::V1::Admin::CategoriesController < Api::V1::Admin::BaseController
   def index
+    authorize Category, :index?
     categories = current_restaurant.categories.ordered
     render json: CategoryBlueprint.render_as_hash(categories)
   end
 
   def create
+    authorize Category, :create?
     category = current_restaurant.categories.build(category_params)
     if category.save
       render json: CategoryBlueprint.render_as_hash(category), status: :created
@@ -15,6 +17,7 @@ class Api::V1::Admin::CategoriesController < Api::V1::Admin::BaseController
 
   def update
     category = current_restaurant.categories.find(params[:id])
+    authorize category
     if category.update(category_params)
       render json: CategoryBlueprint.render_as_hash(category)
     else
@@ -24,6 +27,7 @@ class Api::V1::Admin::CategoriesController < Api::V1::Admin::BaseController
 
   def destroy
     category = current_restaurant.categories.find(params[:id])
+    authorize category
     category.destroy!
     head :no_content
   end
